+++ urbansim_templates/__init__.py
@@ -1,1 +1,1 @@
-version = __version__ = '0.2.dev1'
+version = __version__ = '0.2.dev2'